+++ src/server/config/heatingMode.ts
@@ -1,5 +1,6 @@
 export enum HeatingMode {
   None,
   Winter,
-  Sommer,
+  Summer,
+  TransitionalSeason,
 }